Kill stuck fuse processes
Kill stuck fuse processes if they don't exit after timeout.
Sometimes we observe s3fs processes being stuck even if the volumen is not a mountpoint anymore.
As of now the pod would then be stuck forever in terminating state as we would only run in timeout and kubelet will retry to unmount in a loop
Now we just kill the process if we run into the timeout

Signed-off-by: Benjamin Isinger <Benjamin.Isinger@ibm.com>